Rename parser > useTextToPath
